refactor: página da home
