+++ src/main/webapp/css/paginas/home.css
@@ -1,5 +1,3 @@
-@import url(/css/style.css);
-
 .carousel-item {
     height: 60vh;
     position: relative;
